+++ CHANGELOG.md
@@ -7,12 +7,15 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.10.3 - Unreleased
 
+### Added
+* Added `progkw` as argument to `JobPool.as_completed` to control progress reporting
+* Added `progkw` as argument to `ub.download` / `ub.grabdat` to control progress reporting
+
 
 ### Fixed
 * `ubelt.hash_data` now handles non-numeric float values.
 
 ### Changed
-* Added progress bar keyword arguments to `JobPool.as_completed`
 
 * Changed default of `_hashable_sequence` `types` arg from True to False to be
   consistent, but kept existing types=True behavior in hashable extensions. Changes